+++ test/samTextViewer/UtilsTest.java
@@ -350,8 +350,10 @@ public class UtilsTest {
 	}
 	
 	@Test
-	public void canGetBamReadCount(){
-		assertEquals(15098, Utils.getAlignedReadCount(new File("test_data/ds051.actb.bam")));
+	public void canGetBamReadCount() throws IOException{
+		assertEquals(15098, Utils.getAlignedReadCount("test_data/ds051.actb.bam"));
+		// Painfully slow!
+		// assertEquals(6337212, Utils.getAlignedReadCount("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400SplicesRep2V2.bam"));
 	}
 
 	@Test
@@ -407,13 +409,20 @@ public class UtilsTest {
 	public void canTestForTabixIndex() throws IOException{
 		assertTrue(Utils.hasTabixIndex("test_data/test.bedGraph.gz"));
 		assertTrue(! Utils.hasTabixIndex("test_data/test.bedGraph"));
+
+		// See https://github.com/samtools/htsjdk/issues/797
+		// assertTrue(Utils.hasTabixIndex("ftp://ftp.ensembl.org/pub/release-87/gff3/homo_sapiens/Homo_sapiens.GRCh38.87.abinitio.gff3.gz"));
+		
+		// If this file does not exist, put any valid tabix file and its index on Dropbox/Public and use
+		// the dropbox link here.
+		assertTrue(Utils.hasTabixIndex("http://genome.ucsc.edu/goldenPath/help/examples/vcfExample.vcf.gz"));
 	}
 		
 	@Test
 	public void canGetFileTypeFromName(){
 		
 		assertEquals(TrackFormat.BIGWIG,
-		Utils.getFileTypeFromName("/Users/berald01/Downloads/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig"));
+		Utils.getFileTypeFromName("http://foo/bar/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig"));
 	} 
 
 	@Test
@@ -421,11 +430,21 @@ public class UtilsTest {
 		
 		assertEquals("chrM", Utils.initRegionFromFile("test_data/ds051.short.bam"));
 		assertEquals("chr9", Utils.initRegionFromFile("test_data/hg18_var_sample.wig.v2.1.30.tdf"));
-		assertEquals("chr1:10536", Utils.initRegionFromFile("/Users/berald01/Downloads/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig"));
+		assertEquals("chr1:10536", Utils.initRegionFromFile("test_data/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig"));
 		assertEquals("chr1:67208779", Utils.initRegionFromFile("test_data/refSeq.hg19.short.bed"));
 		assertEquals("chr1:8404074", Utils.initRegionFromFile("test_data/refSeq.hg19.short.sort.bed.gz"));
 		assertEquals("chr1:11874", Utils.initRegionFromFile("test_data/hg19_genes_head.gtf.gz"));
 		assertEquals("chr1:564666", Utils.initRegionFromFile("test_data/wgEncodeDukeDnase8988T.fdr01peaks.hg19.bb"));
+		assertEquals("chr1", Utils.initRegionFromFile("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400SplicesRep2V2.bam"));
+		
+		boolean pass= false;
+		try{
+			Utils.initRegionFromFile(
+					"http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig");
+		} catch(InvalidGenomicCoordsException e){
+			pass= true;
+		}
+		assertTrue(pass);
 		
 		// assertTrue(Utils.initRegionFromFile("hg19:refGene").startsWith("chr1:"));
 	}