+++ test/tracks/TrackSetTest.java
@@ -23,6 +23,18 @@ import samTextViewer.Utils;
 
 public class TrackSetTest {
 
+	@Test
+	public void canConstructTrackSetFromURL() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, BamIndexNotFoundException, InvalidRecordException, SQLException{
+
+		GenomicCoords gc= new GenomicCoords("chr7:5565052-5571960", null, null);
+		
+		TrackSet trackSet= new TrackSet();
+		trackSet.addTrackFromSource("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig", gc, null);
+		trackSet.addTrackFromSource("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400JunctionsRep2V3.bigBed", gc, null);
+		trackSet.addTrackFromSource("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400GeneGencV3cRep2V3.gtf.gz", gc, null);
+		trackSet.addTrackFromSource("http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeCaltechRnaSeq/wgEncodeCaltechRnaSeqGm12878R2x75Il400SigRep2V2.bigWig", gc, null);
+		
+	}
 	
 	@Test
 	public void canGetListOfOpenedFiles() throws ClassNotFoundException, IOException, BamIndexNotFoundException, InvalidGenomicCoordsException, InvalidRecordException, SQLException{
@@ -618,6 +630,14 @@ public class TrackSetTest {
 		assertEquals(0, ts.getTrack(t3).getYLimitMin(), 0.001);
 		assertEquals(10, ts.getTrack(t3).getYLimitMax(), 0.001);
 
+		// First reset all
+		cmdInput= "ylim 0 10";
+		ts.setTrackYlimitsForRegex(Utils.tokenize(cmdInput, " "));
+		for(Track tr : ts.getTrackList()){
+			assertEquals(0, tr.getYLimitMin(), 0.001);
+			assertEquals(10, tr.getYLimitMax(), 0.001);
+		}
+		
 	}
 	
 	@Test